Refactor VSM file structure
