+++ package.json
@@ -51,7 +51,9 @@
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
     "react-flow-renderer": "^9.0.4",
+    "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.3",
+    "recharts": "^2.0.8",
     "uuid": "^3.4.0",
     "web-vitals": "^1.0.1"
   },
