+++ src/components/ValueStreamMap/Buttons/IconButtonStyled.js
