+++ src/components/ValueStreamMap/Buttons/ResetButton.js
@@ -3,7 +3,7 @@ import { Delete } from '@material-ui/icons'
 import { makeStyles, useTheme } from '@material-ui/core/styles'
 import React from 'react'
 
-import { useValueStream } from '../../appContext/valueStreamContext'
+import { useValueStream } from '../valueStreamContext'
 
 const useStyles = makeStyles((theme) => ({
   button: { color: theme.palette.secondary.dark },
