+++ src/components/ValueStreamMap/ConnectionLine.js
