+++ src/components/ValueStreamMap/InputDialog/formConfigs.js
