+++ src/components/ValueStreamMap/Sidebar.js
